refactor: Board 관련 패키지를 추가한다
