+++ src/main/java/chess/board/calculator/OrderBy.java
@@ -1,4 +1,4 @@
-package chess.calculator;
+package chess.board.calculator;
 
 public enum OrderBy {
     ASC, DESC
